+++ demisto_sdk/commands/validate/tests/validators_test.py
@@ -3192,7 +3192,25 @@ def test_validate_no_disallowed_terms_in_customer_facing_docs_end_to_end(repo, m
         (
             {"Packs/test/.pack-ignore"},
             "[file:test.yml]\nignore=BA108\n",
-            b"\n",
+            b"\n\n",
+            "",
+            {
+                "Packs/test/Integrations/test/test.yml",
+            },
+        ),
+        (
+            {"Packs/test/.pack-ignore"},
+            "[file:test.yml]\nignore=BA108\n",
+            b"    ",
+            "",
+            {
+                "Packs/test/Integrations/test/test.yml",
+            },
+        ),
+        (
+            {"Packs/test/.pack-ignore"},
+            "[file:test.yml]\nignore=BA108\n",
+            b"    \n   \n   ",
             "",
             {
                 "Packs/test/Integrations/test/test.yml",
@@ -3217,6 +3235,8 @@ def test_get_all_files_edited_in_pack_ignore(
     - Case 4: old .pack-ignore that is empty and current .pack-ignore that was updated with ignored validation
     - Case 5: old .pack-ignore which is not in the remote branch repo, but only exist in the local branch
     - Case 6: old empty .pack-ignore which contains newlines.
+    - Case 7: old empty .pack-ignore which contains only spaces.
+    - Case 8: old empty .pack-ignore which contains only spaces and newlines.
 
     When:
     - Running get_all_files_edited_in_pack_ignore.
@@ -3228,7 +3248,7 @@ def test_get_all_files_edited_in_pack_ignore(
     - Case 3: Should return both file names.
     - Case 4: Ensure the file that was changed in the .pack-ignore is collected
     - Case 5: Ensure the file that was changed in the .pack-ignore is collected from the local repo
-    - Case 6: Ensure the file that was changed in the .pack-ignore is collected
+    - Case 6 + 7 + 8: Ensure the file that was changed in the .pack-ignore is collected
     """
     mocker.patch.object(
         GitUtil,